ch(cleanup): group profile types
