+++ utils/actions/index.ts
@@ -1,7 +1,7 @@
 export * from "./auth";
 export * from "./developer";
-export * from "./education";
-export * from "./workExperience";
+export * from "./profile/education";
+export * from "./profile/workExperience";
 export * from "./posts";
 export * from "./likes";
 export * from "./comments";
